fix(dopt/dopt): force packages with explicit protocols to pull from registry (#2515)
* fix(dopt/dopt): force packages with explicit protocols to pull from registry

pnpm needs to be forced to link to a remote package (from the registry) instead of a local package: we fix this by adding the link-workspace-packages=false to the .npmrc -- this makes it so that packages which dont have "workspace:*" won't be linked to local packages.  this is fine because we already have a lint check which forces to use "workspace:*". 

pnpm prune fails when you have a mixture of remote packages and local packages: to get around this, I used aliases instead -- in all packages where we depend on a remote copy, you'll see @dopt/react-old which points to the remote copy. this ensure that pnpm prune and all downstream steps build correctly.

* fix(@examples/checklist): attempt to fix by using aliases

pnpm fails to prune correctly, so use aliases

* fix(@app/client,@docs/app): point docs to live sdks, point app/client to aliased sdk

fix docker builds for these two large apps

* fix(@examples/*): alias examples to previous version to fix build issues